Minor addition to gitignore. Altered metrics.models.Observation so it could handle float values such as the seconds measurement returned from Google Analytics. We count fractions of a second where I come from.

+++ metrics/models.py
@@ -87,7 +87,7 @@ class Observation(models.Model):
     metric = models.ForeignKey(Metric, related_name="observations")
     from_datetime = models.DateTimeField()
     to_datetime = models.DateTimeField()
-    value = models.IntegerField(blank=True, null=True)
+    value = models.FloatField(blank=True, null=True)
     payload = models.TextField(blank=True)
     
     class Meta:
